requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sqlalchemy==2.0.32
mysql-connector-python==9.0.0
